+++ packages/core/src/components/legend/legend-group/legend-group.i18n.ts
@@ -1,0 +1,16 @@
+import { Resource } from "../../../utils/i18n";
+
+export const translations: Resource = {
+  en: {
+    "dso-legend-group": {
+      edit: "Edit layers",
+      view: "Leave editmode",
+    },
+  },
+  nl: {
+    "dso-legend-group": {
+      edit: "Kaartlagen bewerken",
+      view: "Bewerken verlaten",
+    },
+  },
+};