+++ packages/core/src/components/legend/legend-item/legend-item.i18n.ts
@@ -1,0 +1,22 @@
+import { Resource } from "../../../utils/i18n";
+
+export const translations: Resource = {
+  en: {
+    "dso-legend-item": {
+      delete: "Remove layer",
+      disabledInfo: "More info",
+      options: "Edit layer",
+      makeActive: "Make {{label}} active",
+      makeActiveDefault: "Make active",
+    },
+  },
+  nl: {
+    "dso-legend-item": {
+      delete: "Kaartlaag verwijderen",
+      disabledInfo: "Meer informatie",
+      options: "Kaartlaag bewerken",
+      makeActive: "Maak {{label}} actief",
+      makeActiveDefault: "Maak actief",
+    },
+  },
+};