+++ packages/dso-toolkit/src/components/legend/legend.args.ts
@@ -4,18 +4,47 @@ import { fn } from "storybook/test";
 
 import { argTypeAction, componentArgs } from "../../storybook";
 
-import { Legend, LegendTabItem } from "./legend.models";
+import { Legend, LegendMode, LegendTabItem } from "./legend.models";
 
 export interface LegendArgs {
+  mode: LegendMode;
+  disabled: boolean;
+  disabledMessage: string;
+  active: boolean;
+  activatable: boolean;
+  label: string;
   tabItems: LegendTabItem[];
   dsoContentSwitch: HandlerFunction;
   dsoClose: HandlerFunction;
+  dsoLegendGroupModeChange: HandlerFunction;
+  dsoDelete: HandlerFunction;
 }
 
 export const legendArgTypes: ArgTypes<LegendArgs> = {
+  mode: {
+    control: { type: "select" },
+    options: ["edit", "view"],
+  },
+  disabled: {
+    control: { type: "boolean" },
+  },
+  disabledMessage: {
+    control: { type: "text" },
+  },
+  active: {
+    control: { type: "boolean" },
+  },
+  activatable: {
+    control: { type: "boolean" },
+  },
+  label: {
+    control: { type: "text" },
+  },
   tabItems: argTypeAction(),
   dsoContentSwitch: argTypeAction(),
   dsoClose: argTypeAction(),
+  dsoLegendGroupModeChange: argTypeAction(),
+  dsoDelete: argTypeAction(),
 };
 
 export const legendaTabItem: LegendTabItem = {
@@ -31,9 +60,17 @@ export const kaartlagenTabItem: LegendTabItem = {
 };
 
 export const legendArgs = componentArgs<LegendArgs>({
+  mode: "view",
+  disabled: false,
+  disabledMessage: "Voorbeeld tekst",
+  active: true,
+  activatable: true,
+  label: "Topografie (BRT)",
   tabItems: [{ ...legendaTabItem, active: true }, kaartlagenTabItem],
   dsoContentSwitch: fn(),
   dsoClose: fn(),
+  dsoLegendGroupModeChange: fn(),
+  dsoDelete: fn(),
 });
 
 export function legendArgsMapper<TemplateFnReturnType>(
@@ -41,7 +78,6 @@ export function legendArgsMapper<TemplateFnReturnType>(
   content: TemplateFnReturnType,
 ): Legend<TemplateFnReturnType> {
   return {
-    ...a,
     tabItems: a.tabItems,
     content,
     dsoContentSwitch: (e) => a.dsoContentSwitch(e.detail),