+++ packages/dso-toolkit/src/components/legend/legend.models.ts
@@ -5,6 +5,8 @@ export interface Legend<TemplateFnReturnType> {
   dsoClose?: (e: CustomEvent<LegendCloseEvent>) => void;
 }
 
+export type LegendMode = "edit" | "view";
+
 export interface LegendTabItem {
   label: string;
   id: string;
@@ -19,4 +21,40 @@ export interface LegendContentSwitchEvent {
 
 export interface LegendCloseEvent {
   originalEvent?: MouseEvent | Event;
+}
+
+export interface LegendItem<TemplateFnReturnType> {
+  options?: TemplateFnReturnType;
+  content?: TemplateFnReturnType;
+  disabled?: boolean;
+  disabledMessage?: string;
+  dsoMouseEnter?: (e: CustomEvent<MouseEvent>) => void;
+  dsoMouseLeave?: (e: CustomEvent<MouseEvent>) => void;
+  dsoActiveChange?: (e: CustomEvent<LegendItemActiveChangeEvent>) => void;
+  dsoDelete?: (e: CustomEvent) => void;
+  active?: boolean;
+  activatable?: boolean;
+  symbol?: TemplateFnReturnType;
+}
+
+export interface LegendItemActiveChangeEvent {
+  /**
+   * De huidige status van de legend.
+   */
+  current: boolean;
+
+  /**
+   * De gewenste status van de legend.
+   */
+  next: boolean;
+
+  originalEvent: Event;
+}
+
+export interface LegendGroup<TemplateFnReturnType = unknown> {
+  mode?: LegendMode;
+  heading?: TemplateFnReturnType;
+  options?: TemplateFnReturnType;
+  children?: TemplateFnReturnType;
+  dsoLegendGroupModeChange?: (e: CustomEvent) => void;
 }