+++ storybook/src/components/legend/legend.core-stories.ts
@@ -1,14 +1,16 @@
-import readme from "@dso-toolkit/core/src/components/legend/readme.md?raw";
+import legendGroupReadme from "@dso-toolkit/core/src/components/legend/legend-group/readme.md?raw";
+import legendItemReadme from "@dso-toolkit/core/src/components/legend/legend-item/readme.md?raw";
+import legendReadme from "@dso-toolkit/core/src/components/legend/readme.md?raw";
 import type { Meta } from "@storybook/web-components-vite";
 import { LegendArgs, legendMeta, legendStories } from "dso-toolkit";
 
 import { templateContainer } from "../../templates";
-import { decorator } from "../legend-item/legend-item.decorator";
 
+import { decorator } from "./legend-item/legend-item.decorator";
 import { kaartlagenRichContent, legendaRichContent } from "./legend.content";
 
 const meta: Meta<LegendArgs> = {
-  ...legendMeta({ readme }),
+  ...legendMeta({ readme: `${legendReadme}\n${legendGroupReadme}\n${legendItemReadme}` }),
   title: "Core/Legend",
 };
 
@@ -21,8 +23,8 @@ const { Legenda, Kaartlagen } = legendStories({
 
     return {
       legendTemplate,
-      legendaRichContent: legendaRichContent(templates),
-      kaartlagenRichContent: kaartlagenRichContent(templates),
+      legendaRichContent,
+      kaartlagenRichContent,
     };
   },
   decorator,